+++ example/demo.ts
@@ -1,11 +1,9 @@
 import Setmore from '../build/main';
 
-const setmoreClient = new Setmore();
-
 // To get your refresh token: https://setmore.docs.apiary.io/#introduction/request-access-to-the-api
-setmoreClient.setRefreshToken(process.env.SETMORE_REFRESH_TOKEN)
+const setmoreClient = new Setmore(process.env.SETMORE_REFRESH_TOKEN);
 
-setmoreClient.staff.get({})
+setmoreClient.staff.get()
 // tslint:disable-next-line: no-console
 .then( value => console.log(value.data))
 // tslint:disable-next-line: no-console