+++ src/lib/auth.ts
@@ -4,7 +4,7 @@ import BaseApi, { ResourcePath } from "./base";
 export default class Staff extends BaseApi {
   public readonly RESOURCE_PATH: ResourcePath = '/o/oauth2/token';
 
-  constructor() {
-    super()
+  constructor(apiClient) {
+    super(apiClient)
   }
 }