+++ src/lib/service.ts
@@ -3,7 +3,7 @@ import BaseApi from "./base";
 export default class Service extends BaseApi {
   public readonly RESOURCE_PATH = '/bookingapi/services';
 
-  constructor() {
-    super()
+  constructor(apiClient) {
+    super(apiClient)
   }
 }