+++ src/lib/setmore.ts
@@ -1,34 +1,30 @@
 import Appointment from './appointment';
-import Auth from './auth';
 import Customer from './customer';
 import Service from './service';
 import Staff from './staff';
 import Timeslot from './timeslot';
-import BaseApi from './base';
+import ApiClient from './apiClient';
+import { AxiosInstance } from 'axios';
 
 class Setmore {
+  public readonly BASE_URI = 'https://developer.setmore.com/api/v1';
   public staff: Staff | undefined;
   public appointments: Appointment | undefined;
   public services: Service | undefined;
   public customers: Customer | undefined;
   public timeslots: Timeslot | undefined;
-  public auth: Auth;
+  public apiClient: AxiosInstance;
 
-  private apis: BaseApi[];
+  constructor(refreshToken) {
+    this.apiClient = ApiClient(this.BASE_URI, refreshToken)
 
-  constructor() {
-    this.staff = new Staff();
-    this.appointments = new Appointment();
-    this.customers = new Customer();
-    this.timeslots = new Timeslot();
-    this.services = new Service();
-    this.apis = [this.staff, this.appointments, this.customers, this.timeslots, this.services]
+    this.staff = new Staff(this.apiClient);
+    this.appointments = new Appointment(this.apiClient);
+    this.customers = new Customer(this.apiClient);
+    this.timeslots = new Timeslot(this.apiClient);
+    this.services = new Service(this.apiClient);
+
   }
-
-  public setRefreshToken (token:string): void {
-    this.apis.forEach( api => api.setRefreshToken(token))
-  }
-
 }
 
 export default Setmore;