+++ src/lib/staff.ts
@@ -4,15 +4,7 @@ import BaseApi, { ResourcePath } from "./base";
 export default class Staff extends BaseApi {
   public readonly RESOURCE_PATH: ResourcePath = '/bookingapi/staffs';
 
-  constructor() {
-    super()
+  constructor(apiClient) {
+    super(apiClient)
   }
-
-  // public get(): Promise<ISetmoreResponse> {
-  //   const config: AxiosRequestConfig = {
-  //     baseURL: this.baseUri,
-  //     url: this.RESOURCE_PATH
-  //   }
-  //   return this.makeRequest(config)
-  // }
 }